+++ src/zepben/evolve/database/sqlite/diagram/diagram_database_reader.py
@@ -11,7 +11,6 @@ from zepben.evolve.database.sqlite.common.metadata_collection_reader import Meta
 from zepben.evolve.database.sqlite.diagram.diagram_database_tables import DiagramDatabaseTables
 from zepben.evolve.database.sqlite.diagram.diagram_service_reader import DiagramServiceReader
 from zepben.evolve.database.sqlite.tables.table_version import TableVersion
-from zepben.evolve.services.common.meta.metadata_collection import MetadataCollection
 from zepben.evolve.services.diagram.diagrams import DiagramService
 
 
@@ -20,7 +19,6 @@ class DiagramDatabaseReader(BaseDatabaseReader):
     A class for reading the `DiagramService` objects and `MetadataCollection` from our diagram database.
 
     :param connection: The connection to the database.
-    :param metadata: The `MetadataCollection` to populate with metadata from the database.
     :param service: The `DiagramService` to populate with CIM objects from the database.
     :param database_description: The description of the database for logging (e.g. filename).
     """
@@ -28,7 +26,6 @@ class DiagramDatabaseReader(BaseDatabaseReader):
     def __init__(
         self,
         connection: Connection,
-        metadata: MetadataCollection,
         service: DiagramService,
         database_description: str,
         tables: DiagramDatabaseTables = None,
@@ -39,7 +36,7 @@ class DiagramDatabaseReader(BaseDatabaseReader):
         tables = tables if tables is not None else DiagramDatabaseTables()
         super().__init__(
             connection,
-            metadata_reader if metadata_reader is not None else MetadataCollectionReader(metadata, tables, connection),
+            metadata_reader if metadata_reader is not None else MetadataCollectionReader(service, tables, connection),
             service_reader if service_reader is not None else DiagramServiceReader(service, tables, connection),
             service,
             database_description,